test multiple table v2

+++ MyDB.php
@@ -41,8 +41,16 @@ EOF;
    $base->exec($query);
    $base->exec($query1);
    var_dump($base);
+   $base->close();
    
-   
+   if(!class_exists('SQLite3'))
+	die('connection failed');
+
+$base = new SQLite3($dbname);
+if(!$base){
+	echo $base->LastErrorMsg();
+}
+else{
 	$base->exec('DROP TABLE IF EXISTS user');
 	
 	$sql =<<<EOF
@@ -62,8 +70,7 @@ EOF;
    else {
       echo "Table user created successfully\n";
    }
-   
-    $base->close();
+   $base->close();
 }
    /*class MyDB extends SQLite3
    {
